Add test for pick&place demo, refactor simulation test script
