+++ tables_demo_planning/src/tables_demo_planning/refactor/tables_demo.py
@@ -118,38 +118,6 @@ class TablesDemoDomain(Domain):
         )
         return problem
 
-    def set_goals(self, problem: Problem, demo_items: List[Item], target_location: Location) -> None:
-        """Set the goals for the overall demo."""
-        problem.clear_goals()
-        assert self.believe_item_at in problem.fluents
-        if any(item.name == "klt_1" for item in demo_items):
-            assert problem.object("in_klt")
-            if any(item.name == "multimeter_1" for item in demo_items):
-                assert problem.object("multimeter_1")
-                problem.add_goal(self.believe_item_at(self.get(Item, "multimeter_1"), self.get(Location, "in_klt")))
-            if any(item.name == "relay_1" for item in demo_items):
-                assert problem.object("relay_1")
-                problem.add_goal(self.believe_item_at(self.get(Item, "relay_1"), self.get(Location, "in_klt")))
-            if any(item.name == "screwdriver_1" for item in demo_items):
-                assert problem.object("screwdriver_1")
-                problem.add_goal(self.believe_item_at(self.get(Item, "screwdriver_1"), self.get(Location, "in_klt")))
-            assert problem.object(target_location.name)
-            problem.add_goal(self.believe_item_at(self.get(Item, "klt_1"), self.objects[target_location.name]))
-        else:
-            if any(item.name == "multimeter_1" for item in demo_items):
-                assert problem.object("multimeter_1")
-                problem.add_goal(
-                    self.believe_item_at(self.get(Item, "multimeter_1"), self.objects[target_location.name])
-                )
-            if any(item.name == "relay_1" for item in demo_items):
-                assert problem.object("relay_1")
-                problem.add_goal(self.believe_item_at(self.get(Item, "relay_1"), self.objects[target_location.name]))
-            if any(item.name == "screwdriver_1" for item in demo_items):
-                assert problem.object("screwdriver_1")
-                problem.add_goal(
-                    self.believe_item_at(self.get(Item, "screwdriver_1"), self.objects[target_location.name])
-                )
-
     def set_search_goals(self, problem: Problem, item_search: Item) -> None:
         """Set the goals for the current item_search subproblem."""
         problem.clear_goals()
